updated projects name
