+++ src/Dbones.Pipes.Tests/Contexts/IConsumer.cs
@@ -1,0 +1,9 @@
+namespace Dbones.Pipes.Tests.Contexts
+{
+    using System.Threading.Tasks;
+
+    public interface IConsumer<in T>
+    {
+        Task Handle(T message);
+    }
+}
