+++ src/Dbones.Pipes.Tests/Contexts/MessageBase.cs
@@ -1,0 +1,10 @@
+namespace Dbones.Pipes.Tests.Contexts
+{
+    using System;
+
+    public abstract class MessageBase
+    {
+        public Guid Id => Guid.NewGuid();
+        public DateTime DateTime => DateTime.UtcNow;
+    }
+}
