+++ src/Dbones.Pipes.Tests/Contexts/OrderPaymentTaken.cs
@@ -1,0 +1,8 @@
+namespace Dbones.Pipes.Tests.Contexts
+{
+    public class OrderPaymentTaken : MessageBase
+    {
+        public decimal Total { get; set; }
+        public string CustomerId { get; set; }
+    }
+}
