+++ src/Dbones.Pipes.Tests/Contexts/OrderPlaced.cs
@@ -1,0 +1,8 @@
+namespace Dbones.Pipes.Tests.Contexts
+{
+    public class OrderPlaced : MessageBase
+    {
+        public decimal Total { get; set; }
+        public string CustomerId { get; set; }
+    }
+}
