+++ src/Dbones.Pipes.Tests/Infrastructure/Injected.cs
@@ -1,0 +1,10 @@
+namespace Dbones.Pipes.Tests.Infrastructure
+{
+    using System;
+
+    public class Injected
+    {
+        public string Value { get; set; }
+        public Type MiddlewareOn { get; set; }
+    }
+}
