+++ src/Dbones.Pipes/PipeItemType.cs
@@ -1,0 +1,8 @@
+﻿namespace Dbones.Pipes
+{
+    public enum PipeItemType
+    {
+        Type,
+        Instance
+    }
+}
