--- src/Pipes.Tests/Contexts/IConsumer.cs
@@ -1,9 +1,0 @@
-namespace Pipes.Tests.Contexts
-{
-    using System.Threading.Tasks;
-
-    public interface IConsumer<in T>
-    {
-        Task Handle(T message);
-    }
-}
