--- src/Pipes.Tests/Contexts/MessageBase.cs
@@ -1,10 +1,0 @@
-namespace Pipes.Tests.Contexts
-{
-    using System;
-
-    public abstract class MessageBase
-    {
-        public Guid Id => Guid.NewGuid();
-        public DateTime DateTime => DateTime.UtcNow;
-    }
-}
