--- src/Pipes.Tests/Contexts/OrderPaymentTaken.cs
@@ -1,8 +1,0 @@
-namespace Pipes.Tests.Contexts
-{
-    public class OrderPaymentTaken : MessageBase
-    {
-        public decimal Total { get; set; }
-        public string CustomerId { get; set; }
-    }
-}
