--- src/Pipes.Tests/Contexts/OrderPlaced.cs
@@ -1,8 +1,0 @@
-namespace Pipes.Tests.Contexts
-{
-    public class OrderPlaced : MessageBase
-    {
-        public decimal Total { get; set; }
-        public string CustomerId { get; set; }
-    }
-}
