--- src/Pipes.Tests/Infrastructure/Injected.cs
@@ -1,10 +1,0 @@
-namespace Pipes.Tests.Infrastructure
-{
-    using System;
-
-    public class Injected
-    {
-        public string Value { get; set; }
-        public Type MiddlewareOn { get; set; }
-    }
-}
