+++ src/Pipes.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31005.135
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Pipes", "Pipes\Pipes.csproj", "{EFE59BDE-9176-4094-921E-DC8C6F5F6C37}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Dbones.Pipes", "Dbones.Pipes\Dbones.Pipes.csproj", "{EFE59BDE-9176-4094-921E-DC8C6F5F6C37}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pipes.Tests", "Pipes.Tests\Pipes.Tests.csproj", "{2FB8E490-EC4C-40B2-BF45-4926B5FEBABF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dbones.Pipes.Tests", "Dbones.Pipes.Tests\Dbones.Pipes.Tests.csproj", "{2FB8E490-EC4C-40B2-BF45-4926B5FEBABF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
