--- src/Pipes/PipeItemType.cs
@@ -1,8 +1,0 @@
-﻿namespace Pipes
-{
-    public enum PipeItemType
-    {
-        Type,
-        Instance
-    }
-}
