--- src/Pipes/Pipes.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageId>Dbones.Pipes</PackageId>
-    <PackageTitle>Dbones.Pipes</PackageTitle>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <PackageTags>Middleware Pipeline</PackageTags>
-    <license>http://opensource.org/licenses/Apache-2.0</license>
-    <PackageProjectUrl>https://github.com/dbones-labs/pipes</PackageProjectUrl>
-    <PackageIconUrl>https://avatars.githubusercontent.com/u/71238176</PackageIconUrl>
-    <RepositoryUrl>https://github.com/dbones-labs/pipes</RepositoryUrl>
-    <IncludeSymbols>true</IncludeSymbols>
-    <IncludeSource>false</IncludeSource>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
-  </ItemGroup>
-
-</Project>
